feat : 견적서 삭제 시 review, payment table에서도 관련 견적서에 해당하는 데이터 삭제

+++ weddingyou_backend/src/main/java/com/mysite/weddingyou_backend/estimate/EstimateService.java
@@ -70,7 +70,7 @@ public class EstimateService {
 	}
 	
 	//견적서 삭제
-	public void delete(int id ) {
+	public void delete(Long id ) {
 		estimateRepository.deleteById(id);
 	}
